requirements/base.txt
Django==1.5.4
South==0.8.2
distribute==0.6.27
dj-static==0.0.5
django-braces==1.2.2
django-crispy-forms==1.4.0
django-extensions==1.2.2
django-model-utils==1.5.0
django-uuslug==1.0.0
python-slugify==0.0.6
gunicorn==0.14.6
psycopg2==2.4.5
static==0.4
wsgiref==0.1.2
